@@ -129,28 +121,43 @@ rendercv = 'rendercv.cli:app'
 # `test` is the virtual environment needed to run the tests of RenderCV.
 
 [tool.hatch.envs.default]
-# Dependencies to be installed in the `default` virtual environment.
+installer = "uv"
+python = "3.13"
 dependencies = [
-    "ruff",    # to lint the code
-    "black",   # to format the code
-    "ipython", # for ipython shell
-    "isort",   # to sort the imports  
-    "pyright", # to check the types
+    "ruff",             # to lint and format the code
+    "black",            # to format the code
+    "ipython",          # for ipython shell
+    "pyright",          # to check the types
+    "pre-commit",       # to run the checks before committing
+    "pytest==8.3.4",    # to run the tests
+    "coverage==7.6.10", # to generate coverage reports
+    "pypdf==5.1.0",     # to read PDF files
 ]
-path = ".venv"
+features = ["full"] # to install full optional dependencies
 [tool.hatch.envs.default.scripts]
 # Hatch allows us to define scripts that can be run in the activated virtual environment
 # with `hatch run ENV_NAME:SCRIPT_NAME`.
 # Format all the code in the `rendercv` package with `black`:
-format = "black rendercv && black docs && black tests" # hatch run format
+format = "black rendercv docs tests && ruff check --fix && ruff format" # hatch run format
 # Lint the code in the `rendercv` package with `ruff`:
-lint = "ruff check rendercv && ruff check tests" # hatch run lint
-# Sort the imports in the `rendercv` package with `isort`:
-sort-imports = "isort rendercv && isort docs && isort tests" # hatch run sort-imports
+lint = "ruff check" # hatch run lint
 # Check types in the `rendercv` package with `pyright`:
-check-types = "pyright rendercv && pyright tests" # hatch run check-types
+check-types = "pyright rendercv tests" # hatch run check-types
+# Run pre-commit checks:
+precommit = "pre-commit run --all-files" # hatch run pre-commit
+# Run the tests:
+test = "pytest" # hatch run test
+# Run the tests and generate the coverage report as HTML:
+test-and-report = "coverage run -m pytest && coverage combine && coverage report && coverage html --show-contexts" # hatch run test-and-report
+
+[tool.hatch.envs.test]
+template = "default"
+[[tool.hatch.envs.test.matrix]]
+python = ["3.10", "3.11", "3.12", "3.13"]
 
 [tool.hatch.envs.docs]
+installer = "uv"
+python = "3.13"
 # Dependencies to be installed in the `docs` virtual environment.
 dependencies = [
     "mkdocs-material==9.5.34",     # to build docs
