@@ -4,7 +4,8 @@ to print nice-looking messages to the terminal.
 """
 
 import functools
-from typing import Callable, Optional
+from collections.abc import Callable
+from typing import Optional
 
 import jinja2
 import pydantic
@@ -14,12 +15,18 @@ import rich.panel
 import rich.progress
 import rich.table
 import rich.text
-import ruamel.yaml
-import ruamel.yaml.parser
-import typer
+
+try:
+    import typer
+except ImportError as e:
+    from .. import _parial_install_error_message
+
+    raise ImportError(_parial_install_error_message) from e
+
+import packaging.version
 from rich import print
 
-from .. import __version__
+from .. import __version__, data
 from . import utilities
 
 
@@ -37,7 +44,9 @@ class LiveProgressReporter(rich.live.Live):
         class TimeElapsedColumn(rich.progress.ProgressColumn):
             def render(self, task: "rich.progress.Task") -> rich.text.Text:
                 elapsed = task.finished_time if task.finished else task.elapsed
-                delta = f"{elapsed:.1f} s"
+                assert elapsed is not None
+                elapsed = elapsed * 1000  # Convert to milliseconds
+                delta = f"{elapsed:.0f} ms"
                 return rich.text.Text(str(delta), style="progress.elapsed")
 
         self.step_progress = rich.progress.Progress(
@@ -119,14 +128,14 @@ def warn_if_new_version_is_available() -> bool:
         True if there is a new version, and False otherwise.
     """
     latest_version = utilities.get_latest_version_number_from_pypi()
-    if latest_version is not None and __version__ != latest_version:
+    version = packaging.version.Version(__version__)
+    if latest_version is not None and version < latest_version:
         warning(
             f"A new version of RenderCV is available! You are using v{__version__},"
             f" and the latest version is v{latest_version}."
         )
         return True
-    else:
-        return False
+    return False
 
 
 def welcome():
@@ -146,13 +155,11 @@ def welcome():
 
     table.add_row("[bold]RenderCV App", "https://rendercv.com")
     table.add_row("Documentation", "https://docs.rendercv.com")
-    table.add_row("Source code", "https://github.com/sinaatalay/rendercv/")
-    table.add_row("Bug reports", "https://github.com/sinaatalay/rendercv/issues/")
-    table.add_row("Feature requests", "https://github.com/sinaatalay/rendercv/issues/")
-    table.add_row("Discussions", "https://github.com/sinaatalay/rendercv/discussions/")
-    table.add_row(
-        "RenderCV Pipeline", "https://github.com/sinaatalay/rendercv-pipeline/"
-    )
+    table.add_row("Source code", "https://github.com/rendercv/rendercv/")
+    table.add_row("Bug reports", "https://github.com/rendercv/rendercv/issues/")
+    table.add_row("Feature requests", "https://github.com/rendercv/rendercv/issues/")
+    table.add_row("Discussions", "https://github.com/rendercv/rendercv/discussions/")
+    table.add_row("RenderCV Pipeline", "https://github.com/rendercv/rendercv-pipeline/")
 
     print(table)
 
@@ -189,8 +196,6 @@ def error(text: Optional[str] = None, exception: Optional[Exception] = None):
     else:
         print()
 
-    raise typer.Exit(code=4)
-
 
 def information(text: str):
     """Print an information message to the terminal.
@@ -212,130 +217,7 @@ def print_validation_errors(exception: pydantic.ValidationError):
     Args:
         exception: The Pydantic validation error object.
     """
-    # This dictionary is used to convert the error messages that Pydantic returns to
-    # more user-friendly messages.
-    error_dictionary: dict[str, str] = {
-        "Input should be 'present'": (
-            "This is not a valid date! Please use either YYYY-MM-DD, YYYY-MM, or YYYY"
-            ' format or "present"!'
-        ),
-        "Input should be a valid integer, unable to parse string as an integer": (
-            "This is not a valid date! Please use either YYYY-MM-DD, YYYY-MM, or YYYY"
-            " format!"
-        ),
-        "String should match pattern '\\d{4}-\\d{2}(-\\d{2})?'": (
-            "This is not a valid date! Please use either YYYY-MM-DD, YYYY-MM, or YYYY"
-            " format!"
-        ),
-        "String should match pattern '\\b10\\..*'": (
-            'A DOI prefix should always start with "10.". For example,'
-            ' "10.1109/TASC.2023.3340648".'
-        ),
-        "URL scheme should be 'http' or 'https'": "This is not a valid URL!",
-        "Field required": "This field is required!",
-        "value is not a valid phone number": "This is not a valid phone number!",
-        "month must be in 1..12": "The month must be between 1 and 12!",
-        "day is out of range for month": "The day is out of range for the month!",
-        "Extra inputs are not permitted": (
-            "This field is unknown for this object! Please remove it."
-        ),
-        "Input should be a valid string": "This field should be a string!",
-        "Input should be a valid list": (
-            "This field should contain a list of items but it doesn't!"
-        ),
-    }
-
-    unwanted_texts = ["value is not a valid email address: ", "Value error, "]
-
-    # Check if this is a section error. If it is, we need to handle it differently.
-    # This is needed because how dm.validate_section_input function raises an exception.
-    # This is done to tell the user which which EntryType RenderCV excepts to see.
-    errors = exception.errors()
-    for error_object in errors.copy():
-        if (
-            "There are problems with the entries." in error_object["msg"]
-            and "ctx" in error_object
-        ):
-            location = error_object["loc"]
-            ctx_object = error_object["ctx"]
-            if "error" in ctx_object:
-                error_object = ctx_object["error"]
-                if hasattr(error_object, "__cause__"):
-                    cause_object = error_object.__cause__
-                    cause_object_errors = cause_object.errors()
-                    for cause_error_object in cause_object_errors:
-                        # we use [1:] to avoid `entries` location. It is a location for
-                        # RenderCV's own data model, not the user's data model.
-                        cause_error_object["loc"] = tuple(
-                            list(location) + list(cause_error_object["loc"][1:])
-                        )
-                    errors.extend(cause_object_errors)
-
-    # some locations are not really the locations in the input file, but some
-    # information about the model coming from Pydantic. We need to remove them.
-    # (e.g. avoid stuff like .end_date.literal['present'])
-    unwanted_locations = ["tagged-union", "list", "literal", "int", "constrained-str"]
-    for error_object in errors:
-        location = error_object["loc"]
-        new_location = [str(location_element) for location_element in location]
-        for location_element in location:
-            location_element = str(location_element)
-            for unwanted_location in unwanted_locations:
-                if unwanted_location in location_element:
-                    new_location.remove(location_element)
-        error_object["loc"] = new_location  # type: ignore
-
-    # Parse all the errors and create a new list of errors.
-    new_errors: list[dict[str, str]] = []
-    for error_object in errors:
-        message = error_object["msg"]
-        location = ".".join(error_object["loc"])  # type: ignore
-        input = error_object["input"]
-
-        # Check if this is a custom error message:
-        custom_message, custom_location, custom_input_value = (
-            utilities.get_error_message_and_location_and_value_from_a_custom_error(
-                message
-            )
-        )
-        if custom_message is not None:
-            message = custom_message
-            if custom_location:
-                # If the custom location is not empty, then add it to the location.
-                location = f"{location}.{custom_location}"
-            input = custom_input_value
-
-        # Don't show unwanted texts in the error message:
-        for unwanted_text in unwanted_texts:
-            message = message.replace(unwanted_text, "")
-
-        # Convert the error message to a more user-friendly message if it's in the
-        # error_dictionary:
-        if message in error_dictionary:
-            message = error_dictionary[message]
-
-        # Special case for end_date because Pydantic returns multiple end_date errors
-        # since it has multiple valid formats:
-        if "end_date" in location:
-            message = (
-                "This is not a valid end date! Please use either YYYY-MM-DD, YYYY-MM,"
-                ' or YYYY format or "present"!'
-            )
-
-        # If the input is a dictionary or a list (the model itself fails to validate),
-        # then don't show the input. It looks confusing and it is not helpful.
-        if isinstance(input, (dict, list)):
-            input = ""
-
-        new_error = {
-            "loc": str(location),
-            "msg": message,
-            "input": str(input),
-        }
-
-        # if new_error is not in new_errors, then add it to new_errors
-        if new_error not in new_errors:
-            new_errors.append(new_error)
+    errors = data.parse_validation_errors(exception)
 
     # Print the errors in a nice table:
     table = rich.table.Table(
@@ -347,19 +229,20 @@ def print_validation_errors(exception: pydantic.ValidationError):
     table.add_column("Input Value", style="magenta")
     table.add_column("Error Message", style="orange4")
 
-    for error_object in new_errors:
+    for error_object in errors:
         table.add_row(
-            error_object["loc"],
+            ".".join(error_object["loc"]),
             error_object["input"],
             error_object["msg"],
         )
 
     print(table)
-    error()  # exit the program
 
 
-def handle_and_print_raised_exceptions(function: Callable) -> Callable:
-    """Return a wrapper function that handles exceptions.
+def handle_and_print_raised_exceptions_without_exit(function: Callable) -> Callable:
+    """Return a wrapper function that handles exceptions. It does not exit the program
+    after an exception is raised. It just prints the error message and continues the
+    execution.
 
     A decorator in Python is a syntactic convenience that allows a Python to interpret
     the code below:
@@ -384,9 +267,16 @@ def handle_and_print_raised_exceptions(function: Callable) -> Callable:
     Returns:
         The wrapped function.
     """
+    try:
+        import ruamel.yaml
+    except Exception as e:
+        from .. import _parial_install_error_message
+
+        raise ImportError(_parial_install_error_message) from e
 
     @functools.wraps(function)
     def wrapper(*args, **kwargs):
+        code = 4
         try:
             function(*args, **kwargs)
         except pydantic.ValidationError as e:
@@ -425,5 +315,49 @@ def handle_and_print_raised_exceptions(function: Callable) -> Callable:
             error(exception=e)
         except Exception as e:
             raise e
+        else:
+            code = 0
+
+        return code
+
+    return wrapper
+
+
+def handle_and_print_raised_exceptions(function: Callable) -> Callable:
+    """Return a wrapper function that handles exceptions. It exits the program after an
+    exception is raised.
+
+    A decorator in Python is a syntactic convenience that allows a Python to interpret
+    the code below:
+
+    ```python
+    @handle_exceptions
+    def my_function():
+        pass
+    ```
+
+    as
+
+    ```python
+    my_function = handle_exceptions(my_function)
+    ```
+
+    which means that the function `my_function` is modified by the `handle_exceptions`.
+
+    Args:
+        function: The function to be wrapped.
+
+    Returns:
+        The wrapped function.
+    """
+
+    @functools.wraps(function)
+    def wrapper(*args, **kwargs):
+        without_exit_wrapper = handle_and_print_raised_exceptions_without_exit(function)
+
+        code = without_exit_wrapper(*args, **kwargs)
+
+        if code != 0:
+            raise typer.Exit(code)
 
     return wrapper
