@@ -4,8 +4,9 @@ field of the input file.
 """
 
 import functools
+import pathlib
 import re
-from typing import Annotated, Any, Literal, Optional, Type, get_args
+from typing import Annotated, Any, Literal, Optional, get_args
 
 import pydantic
 import pydantic_extra_types.phone_numbers as pydantic_phone_numbers
@@ -47,7 +48,7 @@ def validate_url(url: str) -> str:
     return url
 
 
-def create_a_section_validator(entry_type: Type) -> Type[SectionBase]:
+def create_a_section_validator(entry_type: type) -> type[SectionBase]:
     """Create a section model based on the entry type. See [Pydantic's documentation
     about dynamic model
     creation](https://pydantic-docs.helpmanual.io/usage/models/#dynamic-model-creation)
@@ -69,19 +70,17 @@ def create_a_section_validator(entry_type: Type) -> Type[SectionBase]:
         model_name = "SectionWith" + entry_type.__name__.replace("Entry", "Entries")
         entry_type_name = entry_type.__name__
 
-    SectionModel = pydantic.create_model(
+    return pydantic.create_model(
         model_name,
         entry_type=(Literal[entry_type_name], ...),  # type: ignore
         entries=(list[entry_type], ...),
         __base__=SectionBase,
     )
 
-    return SectionModel
-
 
 def get_characteristic_entry_attributes(
-    entry_types: list[Type],
-) -> dict[Type, set[str]]:
+    entry_types: tuple[type],
+) -> dict[type, set[str]]:
     """Get the characteristic attributes of the entry types.
 
     Args:
@@ -98,9 +97,9 @@ def get_characteristic_entry_attributes(
     for EntryType in entry_types:
         all_attributes.extend(EntryType.model_fields.keys())
 
-    common_attributes = set(
+    common_attributes = {
         attribute for attribute in all_attributes if all_attributes.count(attribute) > 1
-    )
+    }
 
     # Store each entry type's characteristic attributes in a dictionary:
     characteristic_entry_attributes = {}
@@ -113,8 +112,8 @@ def get_characteristic_entry_attributes(
 
 
 def get_entry_type_name_and_section_validator(
-    entry: dict[str, str | list[str]] | str | Type, entry_types: list[Type]
-) -> tuple[str, Type[SectionBase]]:
+    entry: dict[str, str | list[str]] | str | type, entry_types: tuple[type]
+) -> tuple[str, type[SectionBase]]:
     """Get the entry type name and the section validator based on the entry.
 
     It takes an entry (as a dictionary or a string) and a list of entry types. Then
@@ -149,7 +148,8 @@ def get_entry_type_name_and_section_validator(
                 break
 
         if entry_type_name is None:
-            raise ValueError("The entry is not provided correctly.")
+            message = "The entry is not provided correctly."
+            raise ValueError(message)
 
     elif isinstance(entry, str):
         # Then it is a TextEntry
@@ -165,7 +165,7 @@ def get_entry_type_name_and_section_validator(
 
 
 def validate_a_section(
-    sections_input: list[Any], entry_types: list[Type]
+    sections_input: list[Any], entry_types: tuple[type]
 ) -> list[entry_types.Entry]:
     """Validate a list of entries (a section) based on the entry types.
 
@@ -199,9 +199,12 @@ def validate_a_section(
                 pass
 
         if entry_type_name is None or section_type is None:
-            raise ValueError(
+            message = (
                 "RenderCV couldn't match this section with any entry types! Please"
-                " check the entries and make sure they are provided correctly.",
+                " check the entries and make sure they are provided correctly."
+            )
+            raise ValueError(
+                message,
                 "",  # This is the location of the error
                 "",  # This is value of the error
             )
@@ -228,10 +231,11 @@ def validate_a_section(
             raise new_error from e
 
     else:
-        raise ValueError(
+        message = (
             "Each section should be a list of entries! Please see the documentation for"
-            " more information about the sections.",
+            " more information about the sections."
         )
+        raise ValueError(message)
     return sections_input
 
 
@@ -247,21 +251,22 @@ def validate_a_social_network_username(username: str, network: str) -> str:
     if network == "Mastodon":
         mastodon_username_pattern = r"@[^@]+@[^@]+"
         if not re.fullmatch(mastodon_username_pattern, username):
-            raise ValueError(
-                'Mastodon username should be in the format "@username@domain"!'
-            )
-    if network == "StackOverflow":
+            message = 'Mastodon username should be in the format "@username@domain"!'
+            raise ValueError(message)
+    elif network == "StackOverflow":
         stackoverflow_username_pattern = r"\d+\/[^\/]+"
         if not re.fullmatch(stackoverflow_username_pattern, username):
-            raise ValueError(
+            message = (
                 'StackOverflow username should be in the format "user_id/username"!'
             )
-    if network == "YouTube":
+            raise ValueError(message)
+    elif network == "YouTube":
         if username.startswith("@"):
-            raise ValueError(
+            message = (
                 'YouTube username should not start with "@"! Remove "@" from the'
                 " beginning of the username."
             )
+            raise ValueError(message)
 
     return username
 
@@ -333,9 +338,7 @@ class SocialNetwork(RenderCVBaseModelWithoutExtraKeys):
 
         network = info.data["network"]
 
-        username = validate_a_social_network_username(username, network)
-
-        return username
+        return validate_a_social_network_username(username, network)
 
     @pydantic.model_validator(mode="after")  # type: ignore
     def check_url(self) -> "SocialNetwork":
@@ -392,6 +395,11 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
         title="Email",
         description="The email address of the person.",
     )
+    photo: Optional[pathlib.Path] = pydantic.Field(
+        default=None,
+        title="Photo",
+        description="Path to the photo of the person, relatie to the input file.",
+    )
     phone: Optional[pydantic_phone_numbers.PhoneNumber] = pydantic.Field(
         default=None,
         title="Phone",
@@ -415,6 +423,22 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
         # `sections` key is preserved for RenderCV's internal use.
         alias="sections",
     )
+
+    @pydantic.field_validator("photo")
+    @classmethod
+    def update_photo_path(cls, value: Optional[pathlib.Path]) -> Optional[pathlib.Path]:
+        """Cast `photo` to Path and make the path absolute"""
+        if value:
+            from .rendercv_data_model import INPUT_FILE_DIRECTORY
+
+            if INPUT_FILE_DIRECTORY is not None:
+                profile_picture_parent_folder = INPUT_FILE_DIRECTORY
+            else:
+                profile_picture_parent_folder = pathlib.Path.cwd()
+
+            return profile_picture_parent_folder / str(value)
+
+        return value
 
     @pydantic.field_validator("name")
     @classmethod
@@ -440,7 +464,7 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
         if self.location is not None:
             connections.append(
                 {
-                    "latex_icon": "\\faMapMarker*",
+                    "typst_icon": "location-dot",
                     "url": None,
                     "clean_url": None,
                     "placeholder": self.location,
@@ -450,7 +474,7 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
         if self.email is not None:
             connections.append(
                 {
-                    "latex_icon": "\\faEnvelope[regular]",
+                    "typst_icon": "envelope",
                     "url": f"mailto:{self.email}",
                     "clean_url": self.email,
                     "placeholder": self.email,
@@ -461,7 +485,7 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
             phone_placeholder = computers.format_phone_number(self.phone)
             connections.append(
                 {
-                    "latex_icon": "\\faPhone*",
+                    "typst_icon": "phone",
                     "url": self.phone,
                     "clean_url": phone_placeholder,
                     "placeholder": phone_placeholder,
@@ -472,7 +496,7 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
             website_placeholder = computers.make_a_url_clean(str(self.website))
             connections.append(
                 {
-                    "latex_icon": "\\faLink",
+                    "typst_icon": "link",
                     "url": str(self.website),
                     "clean_url": website_placeholder,
                     "placeholder": website_placeholder,
@@ -481,22 +505,22 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
 
         if self.social_networks is not None:
             icon_dictionary = {
-                "LinkedIn": "\\faLinkedinIn",
-                "GitHub": "\\faGithub",
-                "GitLab": "\\faGitlab",
-                "Instagram": "\\faInstagram",
-                "Mastodon": "\\faMastodon",
-                "ORCID": "\\faOrcid",
-                "StackOverflow": "\\faStackOverflow",
-                "ResearchGate": "\\faResearchgate",
-                "YouTube": "\\faYoutube",
-                "Google Scholar": "\\faGraduationCap",
-                "Telegram": "\\faTelegram",
+                "LinkedIn": "linkedin",
+                "GitHub": "github",
+                "GitLab": "gitlab",
+                "Instagram": "instagram",
+                "Mastodon": "mastodon",
+                "ORCID": "orcid",
+                "StackOverflow": "stack-overflow",
+                "ResearchGate": "researchgate",
+                "YouTube": "youtube",
+                "Google Scholar": "graduation-cap",
+                "Telegram": "telegram",
             }
             for social_network in self.social_networks:
                 clean_url = computers.make_a_url_clean(social_network.url)
                 connection = {
-                    "latex_icon": icon_dictionary[social_network.network],
+                    "typst_icon": icon_dictionary[social_network.network],
                     "url": social_network.url,
                     "clean_url": clean_url,
                     "placeholder": social_network.username,
@@ -528,7 +552,9 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
 
         if self.sections_input is not None:
             for title, entries in self.sections_input.items():
-                title = computers.dictionary_key_to_proper_section_title(title)
+                formatted_title = computers.dictionary_key_to_proper_section_title(
+                    title
+                )
 
                 # The first entry can be used because all the entries in the section are
                 # already validated with the `validate_a_section` function:
@@ -539,7 +565,7 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
 
                 # SectionBase is used so that entries are not validated again:
                 section = SectionBase(
-                    title=title,
+                    title=formatted_title,
                     entry_type=entry_type_name,
                     entries=entries,
                 )
@@ -550,4 +576,4 @@ class CurriculumVitae(RenderCVBaseModelWithExtraKeys):
 
 # The dictionary below will be overwritten by CurriculumVitae class, which will contain
 # some important data for the CV.
-curriculum_vitae: dict[str, str] = dict()
+curriculum_vitae: dict[str, str] = {}
