@@ -3,17 +3,18 @@
 import copy
 import filecmp
 import itertools
-import os
 import pathlib
 import shutil
 import typing
-from typing import Optional, Type
+import urllib.request
+from typing import Optional
 
 import jinja2
 import pydantic
 import pydantic_extra_types.phone_numbers as pydantic_phone_numbers
 import pypdf
 import pytest
+import ruamel.yaml
 
 from rendercv import data
 from rendercv.renderer import templater
@@ -47,8 +48,8 @@ experience_entry_dictionary = {
     "end_date": "2021-08-12",
     "highlights": [
         (
-            "Developed an [IOS application](https://example.com) that has received"
-            " more than **100,000 downloads**."
+            "Developed an [IOS application](https://example.com) that has received more"
+            " than **100,000 downloads**."
         ),
         "Managed a team of **5** engineers.",
     ],
@@ -170,6 +171,16 @@ def return_a_value_for_a_field_type(
         "start_date": "2015-09",
         "end_date": "2020-06",
         "date": "2021-09",
+        "summary": (
+            "Did *this* and this is a **bold** [link](https://example.com). But I must"
+            " explain to you how all this mistaken idea of denouncing pleasure and"
+            " praising pain was born and I will give you a complete account of the"
+            " system, and expound the actual teachings of the great explorer of the"
+            " truth, the master-builder of human happiness. No one rejects, dislikes,"
+            " or avoids pleasure itself, because it is pleasure, but because those who"
+            " do not know how to pursue pleasure rationally encounter consequences that"
+            " are extremely painful."
+        ),
         "highlights": [
             (
                 "Did *this* and this is a **bold** [link](https://example.com). But I"
@@ -188,7 +199,7 @@ def return_a_value_for_a_field_type(
                 " pleasure."
             ),
         ],
-        "company": "Some **Company**",
+        "company": "Some Company",
         "position": "Software Engineer",
         "name": "My Project",
         "label": "Pro**gram**ming",
@@ -217,7 +228,16 @@ def return_a_value_for_a_field_type(
     field_type_dictionary = {
         pydantic.HttpUrl: "https://example.com",
         pydantic_phone_numbers.PhoneNumber: "+905419999999",
-        str: "A string",
+        str: (
+            "Did *this* and this is a **bold** [link](https://example.com). But I must"
+            " explain to you how all this mistaken idea of denouncing pleasure and"
+            " praising pain was born and I will give you a complete account of the"
+            " system, and expound the actual teachings of the great explorer of the"
+            " truth, the master-builder of human happiness. No one rejects, dislikes,"
+            " or avoids pleasure itself, because it is pleasure, but because those who"
+            " do not know how to pursue pleasure rationally encounter consequences that"
+            " are extremely painful."
+        ),
         list[str]: ["A string", "Another string"],
         int: 1,
         float: 1.0,
@@ -226,20 +246,20 @@ def return_a_value_for_a_field_type(
 
     if field in field_dictionary:
         return field_dictionary[field]
-    elif type(None) in typing.get_args(field_type):
+    if type(None) in typing.get_args(field_type):
         return return_a_value_for_a_field_type(field, field_type.__args__[0])
-    elif typing.get_origin(field_type) == typing.Literal:
+    if typing.get_origin(field_type) == typing.Literal:
         return field_type.__args__[0]
-    elif typing.get_origin(field_type) == typing.Union:
+    if typing.get_origin(field_type) == typing.Union:
         return return_a_value_for_a_field_type(field, field_type.__args__[0])
-    elif field_type in field_type_dictionary:
+    if field_type in field_type_dictionary:
         return field_type_dictionary[field_type]
 
     return "A string"
 
 
 def create_combinations_of_a_model(
-    model: Type[data.Entry],
+    model: type[data.Entry],
 ) -> list[data.Entry]:
     """Look at the required fields and optional fields of a model and create all
     possible combinations of them.
@@ -252,8 +272,8 @@ def create_combinations_of_a_model(
     """
     fields = typing.get_type_hints(model)
 
-    required_fields = dict()
-    optional_fields = dict()
+    required_fields = {}
+    optional_fields = {}
 
     for field, field_type in fields.items():
         value = return_a_value_for_a_field_type(field, field_type)
@@ -278,15 +298,22 @@ def create_combinations_of_a_model(
 
 @pytest.fixture
 def rendercv_filled_curriculum_vitae_data_model(
-    text_entry, bullet_entry
+    text_entry, bullet_entry, testdata_directory_path
 ) -> data.CurriculumVitae:
     """Return a filled CurriculumVitae data model, where each section has all possible
     combinations of entry types.
     """
+    profile_picture_path = testdata_directory_path / "profile_picture.jpg"
+    if update_testdata:
+        # Get an image from https://picsum.photos
+        response = urllib.request.urlopen("https://picsum.photos/id/237/300/300")
+        profile_picture_path.write_bytes(response.read())
+
     return data.CurriculumVitae(
         name="John Doe",
         location="Istanbul, Turkey",
         email="john_doe@example.com",
+        photo=profile_picture_path,  # type: ignore
         phone="+905419999999",  # type: ignore
         website="https://example.com",  # type: ignore
         social_networks=[
@@ -322,7 +349,7 @@ def rendercv_filled_curriculum_vitae_data_model(
 @pytest.fixture
 def jinja2_environment() -> jinja2.Environment:
     """Return a Jinja2 environment."""
-    return templater.setup_jinja2_environment()
+    return templater.Jinja2Environment().environment
 
 
 @pytest.fixture
@@ -373,10 +400,8 @@ def are_these_two_directories_the_same(
         if file1.is_dir():
             if not file2.is_dir():
                 return False
-            are_these_two_directories_the_same(file1, file2)
-        else:
-            if are_these_two_files_the_same(file1, file2) is False:
-                return False
+            return are_these_two_directories_the_same(file1, file2)
+        return are_these_two_files_the_same(file1, file2)
 
     return True
 
@@ -400,16 +425,16 @@ def are_these_two_files_the_same(file1: pathlib.Path, file2: pathlib.Path) -> bo
     if extension1 == ".pdf":
         pages1 = pypdf.PdfReader(file1).pages
         pages2 = pypdf.PdfReader(file2).pages
-        if len(pages1) != len(pages2):
-            return False
+        result = len(pages1) == len(pages2)
 
         for i in range(len(pages1)):
             if pages1[i].extract_text() != pages2[i].extract_text():
-                return False
+                result = False
+                break
 
-        return True
-    else:
-        return filecmp.cmp(file1, file2)
+        return result
+
+    return filecmp.cmp(file1, file2)
 
 
 @pytest.fixture
@@ -457,23 +482,81 @@ def run_a_function_and_check_if_output_is_the_same_as_reference(
                     shutil.move(output_file_path, reference_file_or_directory_path)  # type: ignore
                 else:
                     shutil.move(tmp_path, reference_file_or_directory_path)
-                    os.mkdir(tmp_path)
+                    pathlib.Path.mkdir(tmp_path)
 
         function(tmp_path, reference_file_or_directory_path, **kwargs)
 
         if output_is_a_single_file:
             return are_these_two_files_the_same(
-                output_file_path, reference_file_or_directory_path  # type: ignore
+                output_file_path,  # type: ignore
+                reference_file_or_directory_path,  # type: ignore
             )
-        else:
-            return are_these_two_directories_the_same(
-                tmp_path, reference_file_or_directory_path
-            )
+        return are_these_two_directories_the_same(
+            tmp_path, reference_file_or_directory_path
+        )
 
     return function
 
 
 @pytest.fixture
-def input_file_path(testdata_directory_path) -> pathlib.Path:
+def input_file_path(tmp_path, testdata_directory_path) -> pathlib.Path:
     """Return the path to the input file."""
-    return testdata_directory_path / "John_Doe_CV.yaml"
+    # copy the input file to the temporary directory
+    input_file_path = testdata_directory_path / "John_Doe_CV.yaml"
+    # Update the auxiliary files if update_testdata is True
+    if update_testdata:
+        # create testdata directory if it doesn't exist
+        if not input_file_path.parent.exists():
+            input_file_path.parent.mkdir()
+
+        input_dictionary = {
+            "cv": {
+                "name": "John Doe",
+                "sections": {"test_section": ["this is a text entry."]},
+            },
+        }
+
+        # dump the dictionary to a yaml file
+        yaml_object = ruamel.yaml.YAML()
+        yaml_object.dump(input_dictionary, input_file_path)
+
+    shutil.copyfile(input_file_path, tmp_path / "John_Doe_CV.yaml")
+    return tmp_path / "John_Doe_CV.yaml"
+
+
+@pytest.fixture
+def design_file_path(tmp_path, testdata_directory_path) -> pathlib.Path:
+    """Return the path to the input file."""
+    design_settings_file_path = testdata_directory_path / "John_Doe_CV_design.yaml"
+    if update_testdata:
+        design_settings_file_path.write_text("design:\n  theme: classic\n")
+
+    shutil.copyfile(design_settings_file_path, tmp_path / "John_Doe_CV_design.yaml")
+    return tmp_path / "John_Doe_CV_design.yaml"
+
+
+@pytest.fixture
+def locale_file_path(tmp_path, testdata_directory_path) -> pathlib.Path:
+    """Return the path to the input file."""
+    locale_file_path = testdata_directory_path / "John_Doe_CV_locale.yaml"
+    if update_testdata:
+        locale_file_path.write_text("locale:\n  years: yil\n")
+    shutil.copyfile(locale_file_path, tmp_path / "John_Doe_CV_locale.yaml")
+    return tmp_path / "John_Doe_CV_locale.yaml"
+
+
+@pytest.fixture
+def rendercv_settings_file_path(tmp_path, testdata_directory_path) -> pathlib.Path:
+    """Return the path to the input file."""
+    rendercv_settings_file_path = (
+        testdata_directory_path / "John_Doe_CV_rendercv_settings.yaml"
+    )
+    if update_testdata:
+        rendercv_settings_file_path.write_text(
+            "rendercv_settings:\n  render_command:\n    dont_generate_html: true\n"
+        )
+
+    shutil.copyfile(
+        rendercv_settings_file_path, tmp_path / "John_Doe_CV_rendercv_settings.yaml"
+    )
+    return tmp_path / "John_Doe_CV_rendercv_settings.yaml"
