@@ -15,8 +15,8 @@ from .commands import (
 
 __all__ = [
     "app",
-    "cli_command_render",
-    "cli_command_new",
     "cli_command_create_theme",
+    "cli_command_new",
     "cli_command_no_args",
+    "cli_command_render",
 ]
