@@ -3,6 +3,7 @@ The `rendercv.data.models.rendercv_data_model` module contains the `RenderCVData
 data model, which is the main data model that defines the whole input file structure.
 """
 
+import pathlib
 from typing import Optional
 
 import pydantic
@@ -11,13 +12,18 @@ from ...themes import ClassicThemeOptions
 from .base import RenderCVBaseModelWithoutExtraKeys
 from .curriculum_vitae import CurriculumVitae
 from .design import RenderCVDesign
-from .locale_catalog import LocaleCatalog
+from .locale import Locale
 from .rendercv_settings import RenderCVSettings
+
+INPUT_FILE_DIRECTORY: Optional[pathlib.Path] = None
 
 
 class RenderCVDataModel(RenderCVBaseModelWithoutExtraKeys):
     """This class binds both the CV and the design information together."""
 
+    # `cv` is normally required, but don't enforce it in JSON Schema to allow
+    # `design` or `locale` fields to have individual YAML files.
+    model_config = pydantic.ConfigDict(json_schema_extra={"required": []})
     cv: CurriculumVitae = pydantic.Field(
         title="Curriculum Vitae",
         description="The data of the CV.",
@@ -29,27 +35,45 @@ class RenderCVDataModel(RenderCVBaseModelWithoutExtraKeys):
             "The design information of the CV. The default is the classic theme."
         ),
     )
-    locale_catalog: Optional[LocaleCatalog] = pydantic.Field(
-        default=None,
+    locale: Locale = pydantic.Field(
+        default=None,  # type: ignore
         title="Locale Catalog",
         description=(
             "The locale catalog of the CV to allow the support of multiple languages."
         ),
     )
-    rendercv_settings: Optional[RenderCVSettings] = pydantic.Field(
-        default=None,
+    rendercv_settings: RenderCVSettings = pydantic.Field(
+        default=RenderCVSettings(),
         title="RenderCV Settings",
         description="The settings of the RenderCV.",
     )
 
-    @pydantic.field_validator("locale_catalog")
+    @pydantic.model_validator(mode="before")
     @classmethod
-    def initialize_locale_catalog(
-        cls, locale_catalog: Optional[LocaleCatalog]
-    ) -> Optional[LocaleCatalog]:
-        """Even if the locale catalog is not provided, initialize it with the default
-        values."""
-        if locale_catalog is None:
-            LocaleCatalog()
+    def update_paths(
+        cls, model, info: pydantic.ValidationInfo
+    ) -> Optional[RenderCVSettings]:
+        """Update the paths in the RenderCV settings."""
+        global INPUT_FILE_DIRECTORY  # NOQA: PLW0603
 
-        return locale_catalog
+        context = info.context
+        if context:
+            input_file_directory = context.get("input_file_directory", None)
+            INPUT_FILE_DIRECTORY = input_file_directory
+        else:
+            INPUT_FILE_DIRECTORY = None
+
+        return model
+
+    @pydantic.field_validator("locale", mode="before")
+    @classmethod
+    def update_locale(cls, value) -> Locale:
+        """Update the output folder name in the RenderCV settings."""
+        # Somehow, we need this for `test_if_local_catalog_resets` to pass.
+        if value is None:
+            return Locale()
+
+        return value
+
+
+rendercv_data_model_fields = tuple(RenderCVDataModel.model_fields.keys())
