@@ -1,81 +1,163 @@
-from typing import Literal
+from typing import Literal, Optional
 
-import pydantic
+import rendercv.themes.options as o
 
-from rendercv.data.models.base import RenderCVBaseModelWithoutExtraKeys
-from rendercv.themes.common_options import LaTeXDimension
+o.header_name_bold_field_info.default = False
+o.header_name_font_size_field_info = "25pt"
+o.header_alignment_field_info.default = "left"
 
 
-class ModerncvThemeOptions(RenderCVBaseModelWithoutExtraKeys):
-    """This class is the data model of the theme options for the `moderncv` theme."""
+class Header(o.Header):
+    name_font_size: o.TypstDimension = o.header_name_font_size_field_info
+    name_bold: bool = o.header_name_bold_field_info
+    alignment: o.Alignment = o.header_alignment_field_info
 
-    model_config = pydantic.ConfigDict(extra="forbid")
 
-    theme: Literal["moderncv"]
-    font_size: Literal["10pt", "11pt", "12pt"] = pydantic.Field(
-        default="10pt",
-        title="Font Size",
-        description='The font size of the CV. The default value is "10pt".',
-        examples=["10pt", "11pt", "12pt"],
+o.links_underline_field_info.default = True
+o.links_use_external_link_icon_field_info.default = False
+
+
+class Links(o.Links):
+    underline: bool = o.links_underline_field_info
+    use_external_link_icon: bool = o.links_use_external_link_icon_field_info
+
+
+o.text_font_family_field_info.default = "XCharter"
+o.text_leading_field_info.default = "0.6em"
+
+
+class Text(o.Text):
+    font_family: o.FontFamily = o.text_font_family_field_info
+    leading: o.TypstDimension = o.text_leading_field_info
+
+
+o.section_titles_type_field_info.default = "moderncv"
+o.section_titles_bold_field_info.default = False
+o.section_titles_vertical_space_above_field_info.default = "0.55cm"
+o.section_titles_vertical_space_below_field_info.default = "0.3cm"
+o.section_titles_font_size_field_info.default = "1.4em"
+o.section_titles_line_thickness_field_info.default = "0.15cm"
+
+
+class SectionTitles(o.SectionTitles):
+    type: o.SectionTitleType = o.section_titles_type_field_info
+    vertical_space_above: o.TypstDimension = (
+        o.section_titles_vertical_space_above_field_info
     )
-    page_size: Literal["a4paper", "letterpaper"] = pydantic.Field(
-        default="letterpaper",
-        title="Page Size",
-        description='The page size of the CV. The default value is "letterpaper".',
-        examples=["a4paper", "letterpaper"],
+    vertical_space_below: o.TypstDimension = (
+        o.section_titles_vertical_space_below_field_info
     )
-    color: (
-        Literal["blue"]
-        | Literal["black"]
-        | Literal["burgundy"]
-        | Literal["green"]
-        | Literal["grey"]
-        | Literal["orange"]
-        | Literal["purple"]
-        | Literal["red"]
-    ) = pydantic.Field(
-        default="blue",
-        validate_default=True,
-        title="Primary Color",
-        description='The primary color of the CV. The default value is "blue".',
-        examples=[
-            "blue",
-            "black",
-            "burgundy",
-            "green",
-            "grey",
-            "orange",
-            "purple",
-            "red",
-        ],
+    font_size: o.TypstDimension = o.section_titles_font_size_field_info
+    bold: bool = o.section_titles_bold_field_info
+    line_thickness: o.TypstDimension = o.section_titles_line_thickness_field_info
+
+
+o.entries_vertical_space_between_entries_field_info.default = "0.4cm"
+o.entries_left_and_right_margin_field_info.default = "0cm"
+o.entries_horizontal_space_between_columns_field_info.default = "0.4cm"
+
+
+class Entries(o.Entries):
+    vertical_space_between_entries: o.TypstDimension = (
+        o.entries_vertical_space_between_entries_field_info
     )
-    date_width: LaTeXDimension = pydantic.Field(
-        default="3.8 cm",
-        validate_default=True,
-        title="Date and Location Column Width",
-        description='The width of the date column. The default value is "3.8 cm".',
+    left_and_right_margin: o.TypstDimension = o.entries_left_and_right_margin_field_info
+    horizontal_space_between_columns: o.TypstDimension = (
+        o.entries_horizontal_space_between_columns_field_info
     )
-    content_scale: float = pydantic.Field(
-        default=0.75,
-        title="Content Scale",
-        description=(
-            "The scale of the content with respect to the page size. The default value"
-            ' is "0.75".'
-        ),
+
+
+o.highlights_left_margin_field_info.default = "0cm"
+o.highlights_top_margin_field_info.default = "0.25cm"
+o.highlights_vertical_space_between_highlights_field_info.default = "0.19cm"
+o.highlights_horizontal_space_between_bullet_and_highlight_field_info.default = "0.3em"
+
+
+class Highlights(o.Highlights):
+    left_margin: o.TypstDimension = o.highlights_left_margin_field_info
+    top_margin: o.TypstDimension = o.highlights_top_margin_field_info
+    horizontal_space_between_bullet_and_highlight: o.TypstDimension = (
+        o.highlights_horizontal_space_between_bullet_and_highlight_field_info
     )
-    show_only_years: bool = pydantic.Field(
-        default=False,
-        title="Show Only Years",
-        description=(
-            'If "True", only the years will be shown in the date column. The default'
-            ' value is "False".'
-        ),
+    vertical_space_between_highlights: o.TypstDimension = (
+        o.highlights_vertical_space_between_highlights_field_info
     )
-    disable_page_numbers: bool = pydantic.Field(
-        default=False,
-        title="Disable Page Numbers",
-        description=(
-            'If "True", the page numbers will be disabled. The default value is'
-            ' "False".'
-        ),
+
+
+o.education_entry_main_column_first_row_template_field_info.default = (
+    "**INSTITUTION**, DEGREE in AREA -- LOCATION"
+)
+o.education_entry_degree_column_template_field_info.default = None
+o.entry_base_with_date_date_and_location_column_template_field_info.default = "DATE"
+
+
+class EducationEntry(o.EducationEntry):
+    main_column_first_row_template: str = (
+        o.education_entry_main_column_first_row_template_field_info
     )
+    degree_column_template: Optional[str] = (
+        o.education_entry_degree_column_template_field_info
+    )
+    date_and_location_column_template: str = (
+        o.entry_base_with_date_date_and_location_column_template_field_info
+    )
+
+
+o.normal_entry_main_column_first_row_template_field_info.default = (
+    "**NAME** -- **LOCATION**"
+)
+
+
+class NormalEntry(o.NormalEntry):
+    main_column_first_row_template: str = (
+        o.normal_entry_main_column_first_row_template_field_info
+    )
+    date_and_location_column_template: str = (
+        o.entry_base_with_date_date_and_location_column_template_field_info
+    )
+
+
+o.experience_entry_main_column_first_row_template_field_info.default = (
+    "**POSITION**, COMPANY -- LOCATION"
+)
+
+
+class ExperienceEntry(o.ExperienceEntry):
+    main_column_first_row_template: str = (
+        o.experience_entry_main_column_first_row_template_field_info
+    )
+    date_and_location_column_template: str = (
+        o.entry_base_with_date_date_and_location_column_template_field_info
+    )
+
+
+o.entry_types_education_entry_field_info.default = EducationEntry()
+o.entry_types_normal_entry_field_info.default = NormalEntry()
+o.entry_types_experience_entry_field_info.default = ExperienceEntry()
+
+
+class EntryTypes(o.EntryTypes):
+    education_entry: EducationEntry = o.entry_types_education_entry_field_info
+    normal_entry: NormalEntry = o.entry_types_normal_entry_field_info
+    experience_entry: ExperienceEntry = o.entry_types_experience_entry_field_info
+
+
+o.theme_options_theme_field_info.default = "moderncv"
+o.theme_options_header_field_info.default = Header()
+o.theme_options_text_field_info.default = Text()
+o.theme_options_entry_types_field_info.default = EntryTypes()
+o.theme_options_section_titles_field_info.default = SectionTitles()
+o.theme_options_highlights_field_info.default = Highlights()
+o.theme_options_links_field_info.default = Links()
+o.theme_options_entries_field_info.default = Entries()
+
+
+class ModerncvThemeOptions(o.ThemeOptions):
+    theme: Literal["moderncv"] = o.theme_options_theme_field_info
+    header: Header = o.theme_options_header_field_info
+    highlights: Highlights = o.theme_options_highlights_field_info
+    text: Text = o.theme_options_text_field_info
+    links: Links = o.theme_options_links_field_info
+    entries: Entries = o.theme_options_entries_field_info
+    entry_types: EntryTypes = o.theme_options_entry_types_field_info
+    section_titles: SectionTitles = o.theme_options_section_titles_field_info
